refactor(dashboards): render read-only views from a canvas, not a disabled editor
Extract DashboardGrid/DashboardCanvas so DashboardView can embed blocks without toggling editor preview flags.
