+++ packages/front-end/enterprise/components/Dashboards/DashboardView.tsx
@@ -1,60 +1,46 @@
 import { DashboardInterface } from "shared/enterprise";
-import useApi from "@/hooks/useApi";
-import LoadingOverlay from "@/components/LoadingOverlay";
-import Callout from "@/ui/Callout";
-import DashboardEditor from "@/enterprise/components/Dashboards/DashboardEditor";
+import { Flex } from "@radix-ui/themes";
+import { PiArrowRight } from "react-icons/pi";
+import Link from "@/ui/Link";
 import DashboardSnapshotProvider from "@/enterprise/components/Dashboards/DashboardSnapshotProvider";
+import DashboardCanvas from "@/enterprise/components/Dashboards/DashboardEditor/DashboardCanvas";
+import { getPreviewBlocks } from "@/enterprise/components/Dashboards/DashboardEditor/dashboardLayout";
 
-// Read-only embed: renders a dashboard's blocks without any editing affordances.
-// DashboardEditor only shows its edit entry points when setIsEditing/setBlock
-// are provided, so omitting them here is enough to make this view-only.
 export default function DashboardView({
-  dashboardId,
+  dashboard,
+  maxBlocks,
+  mutate,
 }: {
-  dashboardId: string;
+  dashboard: DashboardInterface;
+  maxBlocks?: number;
+  mutate?: () => void;
 }) {
-  const { data, isLoading, error, mutate } = useApi<{
-    dashboard: DashboardInterface;
-  }>(`/dashboards/${dashboardId}`, {
-    shouldRun: () => Boolean(dashboardId),
-  });
-  const dashboard = data?.dashboard;
-
-  if (isLoading) {
-    return <LoadingOverlay relativePosition />;
-  }
-
-  if (error) {
-    return <Callout status="error">An error occurred: {error.message}</Callout>;
-  }
-
-  if (!dashboard) {
-    return <Callout status="error">Dashboard not found</Callout>;
-  }
+  const visibleBlocks = getPreviewBlocks(dashboard.blocks, maxBlocks);
+  const hasHiddenBlocks = visibleBlocks.length < dashboard.blocks.length;
+  const previewDashboard =
+    visibleBlocks === dashboard.blocks
+      ? dashboard
+      : { ...dashboard, blocks: visibleBlocks };
 
   return (
-    <DashboardSnapshotProvider dashboard={dashboard} mutateDefinitions={mutate}>
-      <DashboardEditor
-        isTabActive
-        id={dashboard.id}
-        initialEditLevel={dashboard.editLevel}
-        ownerId={dashboard.userId}
-        initialShareLevel={dashboard.shareLevel}
-        dashboardOwnerId={dashboard.userId}
-        isGeneralDashboard={true}
-        isEditing={false}
-        title={dashboard.title}
-        blocks={dashboard.blocks}
+    <DashboardSnapshotProvider
+      dashboard={previewDashboard}
+      mutateDefinitions={mutate ?? (() => undefined)}
+    >
+      <DashboardCanvas
+        blocks={visibleBlocks}
         globalControls={dashboard.globalControls}
-        enableAutoUpdates={dashboard.enableAutoUpdates}
-        setBlock={undefined}
-        projects={dashboard.projects ? dashboard.projects : []}
-        mutate={mutate}
-        updateSchedule={dashboard.updateSchedule || undefined}
-        nextUpdate={dashboard.nextUpdate}
-        dashboardLastUpdated={dashboard.lastUpdated}
         dashboardComparison={dashboard.comparison}
+        isGeneralDashboard
+        mutate={mutate}
       />
+      {hasHiddenBlocks && (
+        <Flex justify="end" mt="2">
+          <Link href={`/product-analytics/dashboards/${dashboard.id}`}>
+            View full dashboard <PiArrowRight />
+          </Link>
+        </Flex>
+      )}
     </DashboardSnapshotProvider>
   );
 }
